move stuff to config
