+++ admin/config/beats.py
